refactor(core): introduce AgenticLoopBuilder wrapping the main agentic loop
Move the agentic-loop and single-iteration compositions (topology, workflow
options, and the dowhile continuation predicate) onto an AgenticLoopBuilder
class whose step factories are overridable methods. Step files are untouched
and the existing createAgenticLoopWorkflow / createAgenticExecutionWorkflow
call surfaces are preserved as thin wrappers, so this is behavior-preserving.

This is the skeleton for converging the durable agent loop onto the same
topology: engine variants subclass the builder and override behavior
surgically instead of maintaining a parallel copy of the loop. The params
getter throws when accessed on a builder constructed without main-loop params
so an incompletely-overridden engine subclass fails loudly.
